+++ docs/news.rst
@@ -17,6 +17,7 @@ Downward Lab
   attribute value for all algorithms (Jendrik Seipp).
 * Compute "score_planner_time" and "score_planner_memory" attributes in planner
   parser (Jendrik Seipp).
+* Only consider files ending with ".pddl" and ".sas" when building suites (Jendrik Seipp).
 
 
 v7.0 (2021-10-24)